task#1883: move `TaskCleanupSummary` into `...jobs.models`
