+++ lib/kadai-core/src/main/java/io/kadai/common/internal/KadaiEngineImpl.java
@@ -288,14 +288,14 @@ public class KadaiEngineImpl implements KadaiEngine {
 
   // This should be part of the InternalKadaiEngine. Unfortunately the jobs don't have access to
   // that engine.
-  // Therefore, this getter exits and will be removed as soon as our jobs will be refactored.
+  // Therefore, this getter exists and will be removed as soon as our jobs will be refactored.
   public PriorityServiceManager getPriorityServiceManager() {
     return priorityServiceManager;
   }
 
   // This should be part of the InternalKadaiEngine. Unfortunately the jobs don't have access to
   // that engine.
-  // Therefore, this getter exits and will be removed as soon as our jobs will be refactored.
+  // Therefore, this getter exists and will be removed as soon as our jobs will be refactored.
   public TaskMapper getTaskMapper() {
     return sessionManager.getMapper(TaskMapper.class);
   }
