+++ lib/kadai-core/src/main/java/io/kadai/task/internal/TaskMapper.java
@@ -23,8 +23,8 @@ import io.kadai.common.internal.persistence.MapTypeHandler;
 import io.kadai.common.internal.util.Pair;
 import io.kadai.task.api.CallbackState;
 import io.kadai.task.api.models.TaskSummary;
+import io.kadai.task.internal.jobs.models.TaskCleanupSummary;
 import io.kadai.task.internal.models.MinimalTaskSummary;
-import io.kadai.task.internal.models.TaskCleanupSummary;
 import io.kadai.task.internal.models.TaskImpl;
 import java.time.Instant;
 import java.util.Collection;
