+++ lib/kadai-core/src/main/java/io/kadai/task/internal/jobs/models/TaskCleanupSummary.java
@@ -16,7 +16,7 @@
  *
  */
 
-package io.kadai.task.internal.models;
+package io.kadai.task.internal.jobs.models;
 
 import java.util.Objects;
